Revamp versioning scheme
The website has its own versioning scheme now, and so does the
Windows layout source. If/when more platforms are added, their
layout sources will get their independent version numbers as well.
Tagging scheme (for making GitHub releases) will be as follows:
X.Y.Z, where Z is for the website, Y is for the layout, and X is
for any new platform added.